+++ components/Loading.js
@@ -1,4 +1,5 @@
 import { checkSelector } from '../utils/validations.js'
+import { loadingComponentTemplate } from '../utils/templates.js'
 
 export default function Loading({ selector }) {
   if (new.target !== Loading) return new Loading({ selector })
@@ -6,18 +7,6 @@ export default function Loading({ selector }) {
 
   this.render = () => {
     this.$target = document.querySelector(selector)
-    this.$target.innerHTML = `
-                   <li>
-                    <div class="view">
-                        <label class="label">
-                            <div class="animated-background">
-                                <div class="skel-mask-container">
-                                    <div class="skel-mask"></div>
-                                </div>
-                            </div>
-                        </label>
-                    </div>
-                </li>
-    `
+    this.$target.innerHTML = loadingComponentTemplate
   }
 }
